Add files via upload
Story:
Young Bruce was walking on the streets of Gotham planning to avenge his parents’ death and make Gotham crime-free. He wanted to remain incorruptible.
It was a dark rainy night when he suddenly saw a bat flying. He remembered his fear of bats and got an idea. He decided that he will be disguised and call himself Batman.
Bruce wants you to portray his dark rainy night, so he can keep it as a reminder of the day when he decided to become Batman!